requirements.txt
pymongo==4.14.1
pyTelegramBotAPI==4.29.1
python-dotenv==1.1.1
pytz==2025.2
scrapegraphai==1.62.0
